+++ CHANGES.txt
@@ -1,6 +1,10 @@
 Changes
 -------
 
+0.1.1 (2017-01-16)
+^^^^^^^^^^^^^^^^^^
+* botocore updated to version 1.5.0
+
 0.1.0 (2017-01-12)
 ^^^^^^^^^^^^^^^^^^
 * Pass timeout to aiohttp.request to enforce read_timeout #86 (thanks @vharitonsky)
