+++ aiobotocore/__init__.py
@@ -1,4 +1,4 @@
 from .session import get_session, AioSession
 
-__version__ = '0.1.0'
+__version__ = '0.1.1'
 (get_session, AioSession)  # make pyflakes happy
