+++ aiobotocore/client.py
@@ -1,22 +1,14 @@
 import asyncio
-import copy
 import sys
 
-import botocore.args
-import botocore.auth
 import botocore.client
-import botocore.parsers
-import botocore.serialize
-import botocore.validate
 
-from botocore.exceptions import ClientError, OperationNotPageableError
+from botocore.exceptions import OperationNotPageableError
 from botocore.paginate import Paginator
-from botocore.signers import RequestSigner
 from botocore.utils import get_service_module_name
 
-from .config import AioConfig
-from .endpoint import AioEndpointCreator
 from .paginate import AioPageIterator
+from .args import AioClientArgsCreator
 
 PY_35 = sys.version_info >= (3, 5)
 
@@ -25,10 +17,12 @@ class AioClientCreator(botocore.client.ClientCreator):
 
     def __init__(self, loader, endpoint_resolver, user_agent, event_emitter,
                  retry_handler_factory, retry_config_translator,
-                 response_parser_factory=None, loop=None):
+                 response_parser_factory=None, exceptions_factory=None,
+                 loop=None):
         super().__init__(loader, endpoint_resolver, user_agent, event_emitter,
                          retry_handler_factory, retry_config_translator,
-                         response_parser_factory=response_parser_factory)
+                         response_parser_factory=response_parser_factory,
+                         exceptions_factory=exceptions_factory)
         loop = loop or asyncio.get_event_loop()
         self._loop = loop
 
@@ -39,7 +33,8 @@ class AioClientCreator(botocore.client.ClientCreator):
         # is ClientArgsCreator->AioClientArgsCreator
         args_creator = AioClientArgsCreator(
             self._event_emitter, self._user_agent,
-            self._response_parser_factory, self._loader, loop=self._loop)
+            self._response_parser_factory, self._loader,
+            self._exceptions_factory, loop=self._loop)
         return args_creator.get_client_args(
             service_model, region_name, is_secure, endpoint_url,
             verify, credentials, scoped_config, client_config, endpoint_bridge)
@@ -90,8 +85,10 @@ class AioBaseClient(botocore.client.BaseClient):
             model=operation_model, context=request_context
         )
 
-        if http.status >= 300:
-            raise ClientError(parsed_response, operation_name)
+        if http.status_code >= 300:
+            error_code = parsed_response.get("Error", {}).get("Code")
+            error_class = self.exceptions.from_code(error_code)
+            raise error_class(parsed_response, operation_name)
         else:
             return parsed_response
 
@@ -142,65 +139,3 @@ class AioBaseClient(botocore.client.BaseClient):
         aiohttp does that.
         """
         return self._endpoint._aio_session.close()
-
-
-class AioClientArgsCreator(botocore.args.ClientArgsCreator):
-    def __init__(self, event_emitter, user_agent, response_parser_factory,
-                 loader, loop=None):
-        super().__init__(event_emitter, user_agent,
-                         response_parser_factory, loader)
-        self._loop = loop or asyncio.get_event_loop()
-
-    def get_client_args(self, service_model, region_name, is_secure,
-                        endpoint_url, verify, credentials, scoped_config,
-                        client_config, endpoint_bridge):
-        final_args = self.compute_client_args(
-            service_model, client_config, endpoint_bridge, region_name,
-            endpoint_url, is_secure, scoped_config)
-
-        service_name = final_args['service_name']
-        parameter_validation = final_args['parameter_validation']
-        endpoint_config = final_args['endpoint_config']
-        protocol = final_args['protocol']
-        config_kwargs = final_args['config_kwargs']
-        s3_config = final_args['s3_config']
-        partition = endpoint_config['metadata'].get('partition', None)
-
-        event_emitter = copy.copy(self._event_emitter)
-        signer = RequestSigner(
-            service_name, endpoint_config['signing_region'],
-            endpoint_config['signing_name'],
-            endpoint_config['signature_version'],
-            credentials, event_emitter)
-
-        config_kwargs['s3'] = s3_config
-
-        if isinstance(client_config, AioConfig):
-            connector_args = client_config.connector_args
-        else:
-            connector_args = None
-
-        new_config = AioConfig(connector_args, **config_kwargs)
-        endpoint_creator = AioEndpointCreator(event_emitter, self._loop)
-
-        endpoint = endpoint_creator.create_endpoint(
-            service_model, region_name=endpoint_config['region_name'],
-            endpoint_url=endpoint_config['endpoint_url'], verify=verify,
-            response_parser_factory=self._response_parser_factory,
-            timeout=(new_config.connect_timeout, new_config.read_timeout),
-            connector_args=new_config.connector_args)
-
-        serializer = botocore.serialize.create_serializer(
-            protocol, parameter_validation)
-        response_parser = botocore.parsers.create_parser(protocol)
-        return {
-            'serializer': serializer,
-            'endpoint': endpoint,
-            'response_parser': response_parser,
-            'event_emitter': event_emitter,
-            'request_signer': signer,
-            'service_model': service_model,
-            'loader': self._loader,
-            'client_config': new_config,
-            'partition': partition
-        }
